Exclude various errors to KISS
There are many ways a transaction can fail. Many failures are not
associated with a socket read(2). We can and perhaps should collect and
report error timestamps, but doing so correctly is a lot of work _and_
does not help with monitoring primary SLA-related delays that usually
focus on successful transactions (percentage of unforced errors is
usually a separate metric that is not related to proxy-induced delays).