Update requestReadTimer as soon as possible
Also this change eliminates requestReadTimer special updating
outside ReadFromClient() API.

Also removed requestReadTimer special handling on EOF: in this
case the truncated request is invalid (aborting the transaction?).
For simplicity, we do not maintain requestReadTimer (and others)
if there is a higher level problem (e.g., the entire transaction is
aborted).