Unify *ReadTimer.update() calls
Underlying principle: We want to capture the moment when a successful
transaction(*) received all data it was waiting for -- the time of the
last socket read associated with that transaction.

Some of these EOF events abort the current transaction. We do not care
about such events because we only care about _successful_ transactions.
We (greatly) simplify code by not chasing EOF effects to distinguish
successful EOF events from killer ones (see two previous branch commits
for a negative example of such chasing).

On a client-to-Squid side, a socket read may read data for multiple
(current and/or future) transactions, but at most _one_ of them would be
the current transaction waiting for those bytes. The rest are future
transactions that have not even started yet (from this project events
point of view); a future transaction read timer will _start_ when that
transaction is created, after accumulating enough data to parse its
request header.

(*): Here, "transaction" is a request-response exchange (on either
client-to-Squid or Squid-to-peer connection).