fixup: Time knowledge is optional, but we know what Time is
IME, it is best to use std::optional wrapping to indicate lack of
information rather than to define information type. The difference is
usually very subtle. For example, in this specific case, the only
problematic areas (fixed by this commit) were MessageTimer conversion
constructor (that arguably should take a known time value) and
ExtractRequestStartTime() (that definitely should return a known time
value rather than a never-nil std::optional).

There are exceptions to this rule of thumb, of course. Sometimes, the
possibility of missing information _is_ an intrinsic property of a C++
type. I do not think those exceptions apply to MessageTimer::Time.