Remove timer updates from Http::Tunneler
We claim that "CONNECT tunnel establishment ... happen earlier".
FwdState::successfullyConnectedToPeer() does reset the timer, but
tunnel.cc (which uses Http::Tunneler) does not go through FwdState, and
if feels odd to update a timer that is guaranteed to be reset anyway.

Tunnel establishment handling is one of those areas where "message
activity" (i.e. what we are trying to measure on this branch) differs
from "socket I/O".