Missing error:transaction-end-before-headers in the second transaction
ConnStateData::checkLogging() logged the error if it noticed some bytes
in inBuf, indicating that there is an unparsed yet request. However, in
some cases, the inBuf indication is wrong because inBuf may become empty
when the partially parsed input stored in the parser itself. For
example, this happens for the second (and subsequent) incomplete
requests, when some header lines of the request were received.