Refactored, removing clearPendingRequestBytes()
That method was problematic because it did not clear other
related variables, such as requestErrorBytes and preservedClientData.
With some additional refactoring (also removing checkLogging()), we
do not need this cleanup at all (keeping inBuf and parser intact).

Also fixed %>st for an unparsed transaction. It was invalid if
client sent first line and some header lines. In this case,
the %>st value did not include the first line size.

Also addressed other minor review requests.