Do not require an HTTP request in annotate_client
This ACL needs just the current client-to-Squid connection for
storing the configured annotation: there can be no parsed HTTP request
at that time. If the request is available, it optionally adds the
annotation to the current transaction.

----

Imported Patch217: fix-annotate-client-no-http-request.patch

This patch is based on an unpolished variation of
SQUID-973-annotate-client-should-not-need-a-request commit 0247e16b.
That commit later became a part of master/v7 commit cc71dd7c
"annotate_client and annotate_transaction ACLs must always match".