Explained why importDefaultDirective() looks the way it does
For a minute, I also thought that we should update config_input_line
after processing macros, but then I realized that there are two distinct
contexts here, with two different reporting needs:

* Both this method and processFile() are doing the right thing because
  config_input_line set by them is used for reporting _preprocessing_
  problems, and those problems may be caused by bad macros. They should
  preserve raw input (i.e. input before macro substitution).

* Starting with ApplyPreprocessedConfiguration(), problems can only be
  caused by post-processed input (i.e. input after macro substitution).
  ConfigParser::openDirective() correctly copies post-processed input to
  config_input_line.

N.B. This branch changes config_input_line update timing in
importDefaultDirective() to happen before ProcessMacros(). This change
improves reporting of preprocessing problems. Such problems are
currently very unlikely because default directives do not contain
macros, but not impossible because they do contain the expansion of
`@DEFAULT_...@` sequences, and we do not control the result of that.