fixup: Make initializeOnce() name more unique
When Squid is built with optimizations, this function is usually
inlined. Valgrind does not mangle inlined functions when checking
suppressions, so the original fun:_ZL14initializeOnceiPPc suppression
does not match in optimized builds!

To combat the above problem, I added fun:initializeOnce, a second
initializeOnce() suppression. This one was targeting optimized builds.
After doing that, I began to worry that this function name is too
generic for its essentially "global" suppression context and renamed it
to SquidMainInitializeOnce().

Further testing suggests that dealing with optimized builds is probably
too difficult -- too many suppressions need to be "duplicated" to
accommodate optimization variations. The next change will commit to
non-optimized builds and will remove
initialization-before-primary-loop-optimized suppression added here, but
I wanted to keep this change for the record.

Also polished initializeOnce() description.