fixup: Fixed most (but not all!) proxyProtocolHeaderSent() XXXs
... by mimicking existing FwdState::noteConnection() and
FwdState::tunnelEstablishmentDone() code.

The only aspect we cannot mimic in proxyProtocolHeaderSent() is
Comm::ERR_CLOSING handling. Unfortunately, it is a sign that we should
not inline PROXY protocol header sending code and should create a
dedicated job for writing it instead -- one should not own a connection
and do I/O without a connection closure handler!