fixup: Use Option::dump() to report duplicates
... because that increases our chances of getting value quoting right.
This does not solve all problems because dump() itself does not
reproduce quotes faithfully -- it does not quote values that do not need
quoting. On the other hand, this last bit is not a problem worth solving
because we should rely on generic squid.conf parsing error reporting
code (and not individual parsing functions) to reproduce the exact
directive spelling for triage.