fixup: Removed branch-added ProxyProtocol::Option::quoted_
The quoted_ and related functionality was added when we wanted to
distinguish constant values (unquoted) from logformats (quoted). We
stopped doing that long time ago. There is no reason to maintain this
information, especially since it is not enough to faithfully reproduce
option configuration in the current code -- ConfigParser::QuoteString()
does not quote many values that do not need quoting, even if the admin
did use quotes.