fixup: Avoid `new` in NoNewGlobals for trivially destructable objects
Google C++ Guide: "Objects with static storage duration are forbidden
unless they are trivially destructible".
https://google.github.io/styleguide/cppguide.html#Static_and_Global_Variables

It remains to be seen whether we are allowed this deviation from
NoNewGlobals (or, to be more precise, this NoNewGlobals adjustment), but
it makes sense to me to safely avoid `new` and arguably simplify code.