fixup: Removed branch-added noteFwdPconnUse hacks
noteFwdPconnUse is needed to call PconnPool::noteUses(). That method
counts the number of requests forwarded on a persistent connection (at
connection closing time). ProxyProtocolWriter does not forward requests
and does not reuse idle persistent connections.

We should untangle these stats from persistent connections and (also)
count the number of connections that forwarded no requests (because they
closed too early). However, that improvement is outside this branch
scope, and its implementation is very likely to result in the removal of
noteFwdPconnUse hacks from the rest of Squid code (there are relevant
TODOs about that already)!