Fix cache_peer standby pool destruction
DeleteIdleConnList() casts its void* argument to IdleConnList*.  The
argument contains the hash_link pointer (of the IdleConnList) that had
been stored in the hash_table by hash_join(). The problem is that
hash_join() does not store the original IdleConnList pointer because its
hash_link starts after IdleConnList's vtable. The reverse conversion
from void* to IdleConnList* did not account for that vtable offset.