OT: Maintenance: Make eventAdd(..., true) call less misleading
The last eventAdd() parameter is boolean `cbdata`. That parameter is the
only explicitly boolean parameter, and some other callers use true/false
values to specify whether the event handler data is cbdata-protected. In
this eventAdd() call, the last parameter is not cbdata but event weight!
Most callers use 0 or 1 for weights. Follow that pattern.