Expand %x and %D after bumped SQUID_X509_V_ERR_DOMAIN_MISMATCH
Squid detects SQUID_X509_V_ERR_DOMAIN_MISMATCH errors on various
processing stages, including when receiving an HTTP request on a
successfully bumped TLS connection. If that request targets a domain not
covered by the server certificate, and a mismatch is prohibited, Squid
terminates the transaction with an ERR_SECURE_CONNECT_FAIL response.
That generated error response body lacked %x and %D error details:

```diff
  The system returned:

- [No Error] (TLS code: [Unknown Error Code])
+ [No Error] (TLS code: SQUID_X509_V_ERR_DOMAIN_MISMATCH+broken_cert)

- [No Error Detail]
+ [Certificate does not match domainname: /L=.../O=.../CN=...]
```

The first `[No Error]` expansion of %E remains unchanged because this
particular error does not set `errno`.

TODO: The `broken_cert` part in %x expansion is misleading because Squid
considers that certificate valid and trusted. A mismatch between that
certificate and the new request does not invalidate the former.