Kill SMP disker process if it fails to open its db file
... instead of reporting an error but leaving the disker process idle
forever. After a db opening failure, the disker does not register with
Coordinator and, hence, cannot be used by workers. An unusable process
that failed its primary mission should be killed.

Like any SMP kid, a killed disker will be restarted. Such a restart
might allow the instance to complete startup, but that (small)
probability of recovery is not the primary motivation for this change.
Unsuccessful restarts increase startup noise for a failing instance, but
that (small) drawback is not a valid excuse to avoid this change.