Extend startup timeout to listening socket opening activities
We could use a separate/dedicated timeout, but let's KISS for now.

Also do not rely on FindUnusedHttpSocketSlot() to determine that all
ports have been opened. HttpSockets/NHttpSockets is a hack. The less we
touch or rely on that code, the better. The new code also supports
hypothetical startup activities that are related to opening listening
ports but do not go through clientStartListeningOn() and
clientListenerConnectionOpened() pair of functions. Today, no such
activities are known, but that may change.