fixup: Moved NotifyCoordinator(), addressing branch TODO
The old function decided when to add Strand tag. Today, those decisions
are made at a higher level (and using different tag types), so it makes
sense to adjust the API accordingly.

It is not clear to me whether the function should be located in
StrandJob.cc or StrandKid.cc code, but it is clear that it should not be
inside StrandMessage class code.