Do not collapse on Range requests
Squid is able to collapse range requests, but only some such collapsing
cases are successful. There are three cases to consider here:

* Range-on-Full: A range request collapsing on a regular request.
This case works as expected.

* Range-on-Range: One range request collapsing on another. Squid may
collapse a Range request B on an earlier Range request A but
collapsed request B then usually fails because Range response A, when it
arrives, is marked as not shareable due to its 206 status code.  Squid
does not cache partial responses (i.e. responses with 206 status code).

* Full-on-Range: A full request collapsing on a Range request
This case is similar to the 'Range-on-Range' case above.

Given the above failures/waste, we should disable collapsing on Range
requests in (2) and (3) cases.

The suggested solution ignores the possibility that some Range requests
result in full (i.e. HTTP 200 OK) responses (due to Squid configuration
or due to an origin server decision to ignore Range). Such full
responses would not lead to collapsing failures (they essentially match
the Range-on-Full case ). Still, it looks better to disable all
on-Range collapsing (and miss a few optimization opportunities) than to
keep collapsing and failing in most cases.