Moved clearPublicKeyScope() into StoreEntry::startWriting()
We need to guarantee that clearPublicKeyScope() is called whenever we
write to cache (e.g., from replaceHttpReply()).

Also removed finalReplyIsFresh() precondition. The purpose was to cover
the 'sawDateGoBack' case in http.cc. However, if sawDateGoBack is true,
HttpStateData::reusableReply() returns 'reuseNot' and the entry
becomes private (and we do not need clearPublicKeyScope() for
private entries).